data-preprocessing/requirements.txt
polars==1.35.2
duckdb==1.4.2
pyarrow==22.0.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1
pydantic==2.5.0
pydantic-settings==2.1.0
